Update dynamic-mask.js
Change so that the best fitting mask is actually selected within the dynamic while loop - previously it was checking the length of the value matched against the next mask and comparing its length with the length of the current mask. What it should have been doing is comparing the length of the next mask with the current mask provided the input value matches the next mask.